remove addition of primitive datatypes
Not every language will need primitive reals, angles, and Booleans. Most
will need one or more of them, but rarely will all be necessary. It
should be up to the language designer to add them as needed. It will
also be clearer if the entire grammar is declared in one place, rather
than partially in the application file and partially in the Sasquatch
library.